피드백 반영 수정
- Type enum 객체를 개별 파일에 넣어두고 필요한 곳에 import하는 방식으로 변경
- styled components 파일명 변경
- app.js가 아닌 각 페이지에서 데이터를 fetch하는 것으로 변경
