+++ src/Type.js
@@ -1,0 +1,8 @@
+export const Type = {
+    CATEGORY: "Category",
+    EXHIBITION: "Exhibition",
+    PRODUCT: "Product",
+    BRAND: "Brand"
+};
+
+Object.freeze(Type);
