+++ src/pages/Bookmark/Bookmark.js
@@ -1,14 +1,41 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import Product from "../../components/Product/Product";
-import * as S from "./style";
-const Type = {
-    CATEGORY: "Category",
-    EXHIBITION: "Exhibition",
-    PRODUCT: "Product",
-    BRAND: "Brand"
-};
+import * as S from "./Bookmark.styled";
+import { Type } from "../../Type";
 
-const ProductListPage = ({ bookmarkHandler, bookmarkData }) => {
+
+const ProductListPage = () => {
+    const [bookmarkData, setBookmarkData] = useState([])
+
+    useEffect(() => {
+        const bookmarkData = JSON.parse(localStorage.getItem('bookmark')) || [];
+        setBookmarkData(bookmarkData);
+    }, [])
+    const bookmarkHandler = (product) => {
+        const updatedBookmarkData = [...bookmarkData];
+        const index = updatedBookmarkData.findIndex(item => item.id === product.id);
+
+        if (index === -1) {
+            const obj = {
+                id: product.id,
+                type: product.type,
+                title: product.title,
+                sub_title: product.sub_title,
+                brand_name: product.brand_name,
+                price: product.price,
+                discountPercentage: product.discountPercentage,
+                image_url: product.image_url,
+                brand_image_url: product.brand_image_url,
+                follower: product.follower,
+                isBookmarked: true,
+            };
+            updatedBookmarkData.push(obj);
+        } else {
+            updatedBookmarkData.splice(index, 1);
+        }
+        setBookmarkData(updatedBookmarkData);
+        localStorage.setItem('bookmark', JSON.stringify(updatedBookmarkData));
+    };
 
     const [currentTab, setCurrentTab] = useState(0);
     const data = bookmarkData
