+++ src/pages/Main/Main.js
@@ -1,8 +1,45 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import Product from '../../components/Product/Product';
-import * as S from './style'
+import * as S from './Main.styled'
+import axios from "axios";
 
-const Main = ({ data, bookmarkData, bookmarkHandler }) => {
+const Main = () => {
+    const [data, setData] = useState([])
+    const [bookmarkData, setBookmarkData] = useState([])
+    useEffect(() => {
+        axios.get("http://cozshopping.codestates-seb.link/api/v1/products?count=4")
+            .then(res => setData(res.data))
+            .catch(error => console.log(error))
+
+        const bookmarkData = JSON.parse(localStorage.getItem('bookmark')) || [];
+        setBookmarkData(bookmarkData);
+    }, [])
+
+    const bookmarkHandler = (product) => {
+        const updatedBookmarkData = [...bookmarkData];
+        const index = updatedBookmarkData.findIndex(item => item.id === product.id);
+
+        if (index === -1) {
+            const obj = {
+                id: product.id,
+                type: product.type,
+                title: product.title,
+                sub_title: product.sub_title,
+                brand_name: product.brand_name,
+                price: product.price,
+                discountPercentage: product.discountPercentage,
+                image_url: product.image_url,
+                brand_image_url: product.brand_image_url,
+                follower: product.follower,
+                isBookmarked: true,
+            };
+            updatedBookmarkData.push(obj);
+        } else {
+            updatedBookmarkData.splice(index, 1);
+        }
+        setBookmarkData(updatedBookmarkData);
+        localStorage.setItem('bookmark', JSON.stringify(updatedBookmarkData));
+    };
     return (
         <>
             <S.ProductListWrapper>
