+++ src/pages/Main/Main.styled.js
