+++ src/pages/ProductListPage/ProductListPage.js
@@ -1,16 +1,47 @@
 import React, { useState, useEffect } from "react";
 import Product from "../../components/Product/Product";
 import axios from "axios";
-import * as S from "./style";
-const Type = {
-    CATEGORY: "Category",
-    EXHIBITION: "Exhibition",
-    PRODUCT: "Product",
-    BRAND: "Brand"
-};
+import * as S from "./ProductListPage.styled";
+import { Type } from "../../Type";
 
-const ProductListPage = ({ bookmarkHandler, bookmarkData, data }) => {
+const ProductListPage = () => {
     const [currentTab, setCurrentTab] = useState(0);
+    const [data, setData] = useState([])
+    const [bookmarkData, setBookmarkData] = useState([])
+
+    useEffect(() => {
+        axios.get("http://cozshopping.codestates-seb.link/api/v1/products")
+            .then(res => setData(res.data))
+            .catch(error => console.log(error))
+
+        const bookmarkData = JSON.parse(localStorage.getItem('bookmark')) || [];
+        setBookmarkData(bookmarkData);
+    }, [])
+    const bookmarkHandler = (product) => {
+        const updatedBookmarkData = [...bookmarkData];
+        const index = updatedBookmarkData.findIndex(item => item.id === product.id);
+
+        if (index === -1) {
+            const obj = {
+                id: product.id,
+                type: product.type,
+                title: product.title,
+                sub_title: product.sub_title,
+                brand_name: product.brand_name,
+                price: product.price,
+                discountPercentage: product.discountPercentage,
+                image_url: product.image_url,
+                brand_image_url: product.brand_image_url,
+                follower: product.follower,
+                isBookmarked: true,
+            };
+            updatedBookmarkData.push(obj);
+        } else {
+            updatedBookmarkData.splice(index, 1);
+        }
+        setBookmarkData(updatedBookmarkData);
+        localStorage.setItem('bookmark', JSON.stringify(updatedBookmarkData));
+    };
 
     const all = data && data.map((product) => <Product product={product} key={product.id} bookmarkHandler={bookmarkHandler} isBookmarked={bookmarkData.some(item => item.id === product.id)} />)
     const product = data && (data.filter(data => data.type === Type.PRODUCT)).map((product) => <Product product={product} key={product.id} bookmarkHandler={bookmarkHandler} isBookmarked={bookmarkData.some(item => item.id === product.id)} />)
